Random Number Generator Update
Random number generation is now seeded around the system time such
that it will produce relatively different results. However, if
invocations of ateball are frequent enough, the system time will
not change the seed of the random number generator and the same
result will be given. This is especially true when running ateball
in a script.

TODO:
 * Create a test branch
 * Update random number generator such that it gives different
   results at each invocation of ateball.
 * Employ autoconf over the software.
 * Write help option
 * Write manual page

File(s) Updated:
 * include/answers.h
 * src/answers.c
 * src/main.c