Random Number Generation Update
Updated the seeding of the random number generation. Using system
time as a seed by itself does not make for a quality implementation
for generating random numbers. Running ateball in a script using the
previous implementation generated the same results since the system
time would not change as invocations were so close together.

The current implementation sets a large prime number (31-bit) and
uses the address of said number in combination with its value and
the system time to better generate a 'more random' seed from which
to generate random numbers from. The address of the seed is not
garaunteed to be in the same place every time, and is pseudo-random
in itself.

File(s) Modified:
 * src/main.c